requirements.txt
numpy
pandas
cupy-cuda12x
scikit-learn
